requirements.txt
admin-totals==1.0.1
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ConfigArgParse==0.15.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0
django-admin-rangefilter==0.5.3
django-environ==0.4.4
django-filter==1.0.4
django-grappelli==2.13.2
django-import-export==2.0.1
django-suit==2.0a1
djangorestframework==3.6.4
et-xmlfile==1.0.1
Flask==1.1.1
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
lazy-object-proxy==0.0.0
locustio==0.13.5
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack-python==0.5.6
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.7.3
pylint==2.4.4
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
typed-ast==1.4.0
urllib3==1.25.7
uWSGI==2.0.15
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
